Change site to use Hugo instead of plain html
This uses a custom hugo theme. Because some of the pages are not the
normal blog type pages I've included the html straight from the old page
and just load it rather than using the markdown files. The angular
tutorial is one example of this.

I'm still trying to figure out the best devision between items in the
theme and items at the higher level in the project. Although since this
is only used by this site it likely doesn't matter at all.