Add lazy loading images
This causes some issues since I don't set the height of the image. This causes the screen to jump as the images get loaded in.

I'd need to include a height for each image automatically with hugo so I don't need to specify when referencing the images in the md file.